chore: fix buildings tests
